rename src/renderer/components
