+++ src/renderer/components/stats-display.vue
